Virtual host file generation

+++ fabfile.py
@@ -3,13 +3,14 @@ import os
 import string
 import random
 from fabric.api import run, cd, env, settings, hide, sudo, prompt, local, put
-from fabric.contrib.files import exists
+from fabric.contrib.files import exists, upload_template
 from fabric.utils import warn
 
 env.hosts = ['root@162.243.102.184']
 
 rootdir = "/var/www/"
 homedir = rootdir + "campaign/"
+apache_conf = '/etc/apache2/sites-enabled/campaign.conf'
 
 def install():
     run("apt-get install -y --no-upgrade python-pip build-essential git libmysqlclient-dev apache2 python-dev libapache2-mod-wsgi", shell=False)
@@ -27,8 +28,13 @@ def install_code():
         run("git pull")
         run("virtualenv env --no-site-packages")
         run("source %s/env/bin/activate && pip install -r %s/requirements.txt" % (homedir, homedir))
-        # TODO: copy vhosts file
 
+        if not exists(apache_conf):
+            domain = prompt('Enter domain (e.g. campaign.example.com):')
+            apache_context = {'homedir' : homedir, 'servername' : domain}
+            upload_template(filename='vhost.conf.jinja', destination=apache_conf, 
+                            use_jinja=True, context=apache_context)
+            
     apache_restart()
 
 def install_mysql():
@@ -50,26 +56,31 @@ def install_mysql():
 
 
     # DB set up
-    if root_password is None:
-        root_password = prompt('Please enter MySQL root password:')
-    dbname = prompt('Enter database name:')
-    run_mysql_cmd('root', root_password, "CREATE DATABASE IF NOT EXISTS %s;" % dbname)
-    username = prompt('Enter mysql username:')
-    password = prompt('Enter mysql password:')
-    run_mysql_cmd('root', root_password, 
-                  "GRANT ALL PRIVILEGES ON %s.* TO '%s'@'localhost' IDENTIFIED BY '%s';" % 
-                  (dbname, username, password))
+    db_setup = prompt('Setup database [y/n]?')
+    if db_setup == 'y':
+        if root_password is None:
+            root_password = prompt('Please enter MySQL root password:')
+        dbname = prompt('Enter database name:')
+        run_mysql_cmd('root', root_password, "CREATE DATABASE IF NOT EXISTS %s;" % dbname)
+        username = prompt('Enter mysql username:')
+        password = prompt('Enter mysql password:')
+        run_mysql_cmd('root', root_password, 
+                      "GRANT ALL PRIVILEGES ON %s.* TO '%s'@'localhost' IDENTIFIED BY '%s';" % 
+                      (dbname, username, password))
     
     # Generate secrets file
     secret_path = homedir+"campaign/secret.py"
-    if not exists(secret_path):
+    if db_setup == 'y' and not exists(secret_path):
         secrets = {}
         secrets['tw_sid'] = prompt('Enter Twilio sid:')
         secrets['tw_token'] = prompt('Enter Twilio token:')
         secrets['tw_caller_id'] = prompt('Enter Twilio caller id (e.g. +15555555555):')
         secrets['BASE_URL'] = prompt('Enter base URL e.g. (http://campaign.example.com/:')
         secrets['SECRET_KEY'] = generate_secret_key()
-        
+        secrets['MYSQL_USER'] = username
+        secrets['MYSQL_PASSWORD'] = password
+        secrets['MYSQL_DB'] = dbname
+        secrets['STATIC_ROOT'] = homedir + 'static/'
         temp_path = 'secret.temp'
         f = open(temp_path, 'w')
         for key, value in secrets.iteritems():
@@ -78,7 +89,9 @@ def install_mysql():
         put(temp_path, secret_path)
         os.remove(temp_path)
     # TODO: Sunlight Foundation API key
-    
+
+    # Now go to homedir, export PYTHONPATH=$(pwd); python campaign/manage.py syncdb; python campaign/manage.py migrate core
+
 def generate_secret_key():
     return "".join([random.SystemRandom().choice(string.digits + string.letters + string.punctuation) for i in range(100)])
